third_party/WebKit/LayoutTests/http/tests/htmlimports/import-normalize-213.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/normalize-test-helper.js"></script>
<link rel="import" href="resources/normalize-child-2.html">
<link rel="import" href="resources/normalize-child-1.cgi">
<link rel="import" href="resources/normalize-child-3.html">
</head>
<body>
<script>
test(function() {
    assert_array_equals(importedDocumentList, ['normalize-child-4.html', 'normalize-child-3.html', 'normalize-child-2.html', 'normalize-child-1.cgi']);
}, 'Imports are evaluated in dependency order regardeless of the availability order.');
</script>
</body>
</html>
